airflow/requirements.txt
apache-airflow-providers-postgres
psycopg2-binary
requests
pandas
sqlalchemy
